@@ -3,14 +3,13 @@ source "https://rubygems.org"
 ruby "2.4.0"
 
 # the most important stuff
-gem "rails", "4.2.8"
+gem "rails", "~> 5.0.2"
 gem "pg"
-gem "rails-api"
-gem "rails_api_format", git: "https://github.com/RuslanKhabibullinFs/rails-api-format.git"
 
 # all other gems
 gem "active_model_serializers", git: "https://github.com/rails-api/active_model_serializers.git"
 gem "decent_exposure"
+gem "devise"
 gem "dotenv-rails"
 gem "health_check"
 gem "interactor"
@@ -19,7 +18,6 @@ gem "rack-cors", require: "rack/cors"
 gem "responders"
 gem "rollbar"
 gem "seedbank"
-gem "simple_token_authentication"
 gem "thin"
 gem "jwt"
 
@@ -59,8 +57,4 @@ group :development, :test, :staging do
   gem "factory_girl_rails"
   gem "rspec_api_documentation"
   gem "apitome"
-end
-
-group :staging, :production do
-  gem "rails_12factor"
 end
