@@ -5,8 +5,10 @@
 
 # Enable parameter wrapping for JSON. You can disable this by setting :format to an empty array.
 ActiveSupport.on_load(:action_controller) do
-  include ActionController::ParamsWrapper
-
-  # Enable parameter wrapping for JSON
-  wrap_parameters(format: %i(json url_encoded_form))
+  wrap_parameters format: [:json]
 end
+
+# To enable root element in JSON for ActiveRecord objects.
+# ActiveSupport.on_load(:active_record) do
+#   self.include_root_in_json = true
+# end
