@@ -1,6 +1,4 @@
 FactoryGirl.define do
-  sequence(:username) { |n| "user#{n}" }
+  sequence(:email) { Faker::Internet.email }
   sequence(:password) { "123456" }
-  sequence(:article_title) { Faker::Lorem.sentence }
-  sequence(:article_content) { Faker::Lorem.paragraph }
 end
